店面完成

+++ app/Common/Common.php
@@ -245,11 +245,13 @@ class Common
 
     // 删除多张图片----读取数据图片字段
     public function DataPicDelAll($val){
-        foreach (unserialize($val) as $List){
-            if(!empty($List)){
-                $images = public_path($this->pic_path.$List);
-                if (file_exists ($images )) {
-                    unlink ($images);
+        if (!empty($val)){
+            foreach (unserialize($val) as $List){
+                if(!empty($List)){
+                    $images = public_path($this->pic_path.$List);
+                    if (file_exists ($images )) {
+                        unlink ($images);
+                    }
                 }
             }
         }
